Move ephemeral negotiation to own module
