+++ talpid-tunnel-config-client/src/socket.rs
@@ -12,15 +12,13 @@ mod sys {
 
     pub use libc::{setsockopt, socklen_t, IPPROTO_TCP, TCP_MAXSEG};
     pub use std::os::fd::{AsRawFd, RawFd};
-    use std::pin::Pin;
-    use std::task::{Context, Poll};
-    use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
 
     /// MTU to set on the tunnel config client socket. We want a low value to prevent fragmentation.
     /// Especially on Android, we've found that the real MTU is often lower than the default MTU, and
     /// we cannot lower it further. This causes the outer packets to be dropped. Also, MTU detection
     /// will likely occur after the PQ handshake, so we cannot assume that the MTU is already
     /// correctly configured.
+    /// This is set to the lowest possible IPv4 MTU.
     const CONFIG_CLIENT_MTU: u16 = 576;
 
     pub struct TcpSocket {
